requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
sqlparse==0.5.1
langsmith